WA-238 Rename transactions form component to AddTRansaction
